Migrate activity management out of admin namespace
