+++ app/views/activities/_form.haml
@@ -1,6 +1,6 @@
 = errors_for(@activity)
 
-= form_for [:admin, @activity] do |f|
+= form_for @activity do |f|
   .form-group
     = f.label :name
     = f.text_field :name, placeholder: "Name", class: "form-control"
