+++ app/views/activities/edit.haml
