+++ app/views/activities/new.haml
